V3

+++ commands/clear.js
@@ -25,10 +25,22 @@ async function clearChannel(channelId, client) {
 }
 
 export async function run(message, client) {
-        if (message.content.startsWith('%clear')) {
-            await clearChannel(message.channel.id, client);
-            return message.channel.send('Channel cleared!').then(msg => {
-                setTimeout(() => msg.delete(), 5000);
-            })
-        }
+// Obtener el GuildMember del autor del mensaje
+const guildMember = await message.guild.members.fetch(message.author.id);
+        
+// ID del rol requerido
+const requiredRoleId = '731870690206154785'; // Reemplaza con el ID del rol necesario
+
+// Comprobar si el autor tiene el rol requerido
+if (guildMember.roles.cache.has(requiredRoleId)) {
+    if (message.content.startsWith('%clear')) {
+        await clearChannel(message.channel.id, client);
+        return message.channel.send('Channel cleared!').then(msg => {
+            setTimeout(() => msg.delete(), 5000);
+        })
+    }
+} else {
+    // El usuario no tiene el rol requerido
+    return message.reply('No tienes los permisos necesarios para ejecutar este comando.');
+}
 }
